requirements.txt
PyQt5==5.15.10
opencv-python==4.9.0.80
numpy==1.26.4
Pillow==10.2.0
transformers==4.37.2
torch==2.1.2
torchvision==0.16.2
scikit-image==0.21.0
python-magic==0.4.27
pytesseract==0.3.10
moviepy==1.0.3
faiss-cpu==1.7.4 